+++ lib/src/input/core/valid_input.dart
@@ -5,7 +5,7 @@
 import 'package:meta/meta.dart';
 import 'package:valid_forms/src/core/listenable.dart';
 
-import '../validator/input_validator.dart';
+import '../../validator/core/input_validator.dart';
 
 /// a field to be validated against zero or more validators.
 abstract class ValidInput<T, V> extends Listenable {