+++ lib/src/predicate/core/validation_predicate.dart
@@ -1,0 +1,7 @@
+abstract class ValidationPredicate {
+  const ValidationPredicate();
+
+  bool call();
+}
+
+