+++ lib/src/predicate/custom_validation_predicate.dart
@@ -1,0 +1,11 @@
+import 'core/validation_predicate.dart';
+
+class CustomValidationPredicate extends ValidationPredicate {
+  const CustomValidationPredicate(bool Function() predicate)
+      : _predicate = predicate;
+
+  final bool Function() _predicate;
+
+  @override
+  bool call() => _predicate();
+}