+++ lib/src/validator/custom_input_validator.dart
@@ -1,0 +1,8 @@
+import 'core/input_validator.dart';
+
+class CustomInputValidator extends InputValidator {
+  CustomInputValidator({
+    required super.predicate,
+    required super.invalidReason,
+  });
+}