+++ lib/valid_forms.dart
@@ -4,10 +4,16 @@
 
 library valid_forms;
 
-export 'src/core/valid_form.dart';
-export './src/input/valid_input.dart';
+export './src/constant/phone_number_area.g.dart';
+export './src/core/valid_form.dart';
+export './src/input/core/valid_input.dart';
 export './src/input/confirmable_valid_input.dart';
 export './src/input/valid_text_input.dart';
-export 'src/validator/input_validator.dart';
-export 'src/validator/phone_number/fixed_line_number_validator.dart';
-export 'src/validator/phone_number/mobile_number_validator.dart';
+export './src/predicate/core/validation_predicate.dart';
+export './src/predicate/custom_validation_predicate.dart';
+export './src/predicate/fixed_line_number_validation_predicate.dart';
+export './src/predicate/mobile_number_validation_predicate.dart';
+export './src/validator/core/input_validator.dart';
+export './src/validator/custom_input_validator.dart';
+export './src/validator/fixed_line_number_input_validator.dart';
+export './src/validator/mobile_number_input_validator.dart';