+++ test/validator/input_validator_test.dart
@@ -1,5 +1,6 @@
 import 'package:test/test.dart';
-import 'package:valid_forms/src/validator/input_validator.dart';
+import 'package:valid_forms/src/predicate/custom_validation_predicate.dart';
+import 'package:valid_forms/src/validator/core/input_validator.dart';
 
 class _FakeValidFieldValidator extends InputValidator {
   _FakeValidFieldValidator({
@@ -12,7 +13,7 @@ void main() {
   group('$InputValidator', () {
     test('should return predicate value on validation', () {
       // arrange
-      bool predicate() => false;
+      final predicate = CustomValidationPredicate(() => false);
       const invalidReason = 'value is false';
       final sut = _FakeValidFieldValidator(
         predicate: predicate,